extract out fungible token standard as a separate crate and clone storage management as a separate crate
