+++ examples/fungible-token/ft/src/lib.rs
@@ -15,13 +15,13 @@ NOTES:
   - To prevent the deployed contract from being modified or deleted, it should not have any access
     keys on its account.
 */
-use near_contract_standards::fungible_token::metadata::{
+use near_ft_standard::fungible_token::metadata::{
     FungibleTokenMetadata, FungibleTokenMetadataProvider, FT_METADATA_SPEC,
 };
-use near_contract_standards::fungible_token::{
+use near_ft_standard::fungible_token::{
     FungibleToken, FungibleTokenCore, FungibleTokenResolver,
 };
-use near_contract_standards::storage_management::{
+use near_storage_standard::storage_management::{
     StorageBalance, StorageBalanceBounds, StorageManagement,
 };
 use near_sdk::borsh::{BorshDeserialize, BorshSerialize};
@@ -83,7 +83,7 @@ impl Contract {
         this.token.internal_register_account(&owner_id);
         this.token.internal_deposit(&owner_id, total_supply.into());
 
-        near_contract_standards::fungible_token::events::FtMint {
+        near_ft_standard::fungible_token::events::FtMint {
             owner_id: &owner_id,
             amount: total_supply,
             memo: Some("new tokens are minted"),
