+++ near-contract-standards/Cargo.toml
@@ -15,6 +15,7 @@ NEAR smart contracts standard library.
 [dependencies]
 near-sdk = { path = "../near-sdk", version = "~5.0.0", default-features = false, features = ["legacy"] }
 
+
 [dev-dependencies]
 near-sdk = { path = "../near-sdk", default-features = false, features = ["unit-testing"] }
