+++ near-contract-standards/src/lib.rs
@@ -2,7 +2,7 @@
 #![allow(clippy::missing_const_for_fn, clippy::redundant_pub_crate)]
 
 /// Fungible tokens as described in [by the spec](https://nomicon.io/Standards/FungibleToken/README.html).
-pub mod fungible_token;
+//pub mod fungible_token;
 /// Non-fungible tokens as described in [by the spec](https://nomicon.io/Standards/NonFungibleToken/README.html).
 pub mod non_fungible_token;
 /// Storage management deals with handling [state storage](https://docs.near.org/docs/concepts/storage-staking) on NEAR. This follows the [storage management standard](https://nomicon.io/Standards/StorageManagement.html).
