+++ near-ft-standard/src/event.rs
@@ -8,8 +8,8 @@ use near_sdk::serde_json;
 #[must_use = "don't forget to `.emit()` this event"]
 #[serde(rename_all = "snake_case")]
 pub(crate) enum NearEvent<'a> {
-    Nep171(crate::non_fungible_token::events::Nep171Event<'a>),
-    //Nep141(near_fungible_token::fungible_token::events::Nep141Event<'a>),
+    //Nep171(near_contract_standards::non_fungible_token::events::Nep171Event<'a>),
+    Nep141(crate::fungible_token::events::Nep141Event<'a>),
 }
 
 impl<'a> NearEvent<'a> {
