+++ near-ft-standard/src/fungible_token/macros.rs
