+++ near-ft-standard/src/fungible_token/metadata.rs
