+++ near-ft-standard/src/fungible_token/receiver.rs
