+++ near-ft-standard/src/fungible_token/resolver.rs
