+++ near-ft-standard/src/lib.rs
@@ -1,0 +1,2 @@
+pub mod fungible_token;
+pub(crate) mod event;
