+++ near-storage-standard/Cargo.toml
@@ -1,0 +1,9 @@
+[package]
+name = "near-storage-standard"
+edition = "2021"
+version.workspace = true
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[dependencies]
+near-sdk = { path = "../near-sdk", version = "~5.0.0", default-features = false, features = ["legacy"] }
