+++ near-storage-standard/src/lib.rs
@@ -1,0 +1,1 @@
+pub mod storage_management;
